Move all stuff in main_test.go into utitls_test.go
